requirements.txt
GitPython == 2.1.15
aiogoogletrans == 3.3.2
aiohttp == 3.6.2
async_timeout == 3.0.1
beautifulsoup4==4.9.3
bs4==0.0.1
click == 7.1.2
zidiscord.py == 1.5.1
epicstore_api == 0.0.4
numexpr == 2.7.1
praw == 7.1.0
python_dateutil == 2.8.1
python_dotenv == 0.14.0
pytz == 2020.1
requests_oauthlib == 1.3.0
quart == 0.13.1
requests == 2.24.0
TagScriptEngine == 2.3.0
youtube_dl == 2020.9.20

-e git://github.com/ext-creators/discord-ext-ipc.git#egg=discord.ext.ipc
